requirements.txt
absl-py==0.2.2
alembic==1.0.10
appdirs==1.4.3
apptools==4.4.0
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.3
async-generator==1.10
atomicwrites==1.2.0
attrs==18.1.0
Automat==0.3.0
backcall==0.1.0
bleach==1.5.0
cachetools==4.0.0
catkin-pkg==0.4.12
certifi==2019.11.28
certipy==0.1.3
cffi==1.11.5
characteristic==14.3.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==0.5.5
configobj==5.0.6
constantly==15.1.0
correlation-cuda==0.0.0
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
dask==0.19.0
decorator==4.3.0
descartes==1.0.1
docutils==0.14
easydict==1.9
entrypoints==0.2.3
envisage==4.7.1
ete3==3.1.1
Fiona==1.8.6
freetype-py==2.1.0.post1
future==0.17.0
fuzzyfinder==2.1.0
gast==0.2.0
GPUtil==1.3.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
imageio==2.3.0
incremental==17.5.0
ipykernel==4.8.2
ipympl==0.2.1
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
jupyterhub==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.3
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Mako==1.0.9
Markdown==2.6.11
MarkupSafe==1.0
mathutils==2.78
matplotlib==3.0.1
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
mpld3==0.3
msgpack==0.6.0
msgpack-numpy==0.4.4.2
multipledispatch==0.6.0
munch==2.3.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.6.0
numpy==1.14.5
oauthlib==3.0.1
olefile==0.45.1
open3d==0.6.0.0
opencv-python==3.4.2.17
pamela==1.0.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.7.1
projectq==0.3.6
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
protonets==0.0.1
ptyprocess==0.6.0
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybind11==2.2.4
pybullet==2.5.6
pycocotools==2.0
pycollada==0.6
pycparser==2.18
pyface==6.0.0
pyglet==1.4.0b1
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==1.8.4
PyOpenGL==3.1.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyquaternion==0.9.5
pyrender==0.1.32
pyrr==0.10.1
pytest==3.7.3
python-dateutil==2.7.3
python-editor==1.0.4
python-prctl==1.7
pytz==2018.5
PyWavefront==0.2.0
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.20.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.3.3
statsmodels==0.9.0
tabulate==0.8.2
tensorboard==1.10.0
tensorboardX==1.5
tensorflow-gpu==1.10.1
tensorpack==0.9.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.3.0
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
traits==5.1.1
traitsui==6.0.0
trimesh==3.4.0
Twisted==18.7.0
typing==3.6.4
urdfdom-py==0.4.0
urllib3==1.24
visdom==0.1.8.8
vtk==8.1.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wrapt==1.10.11
xmltodict==0.11.0
yacs==0.1.6
